hwInit (audio) 98.78->100: #14 unsigned-cast on salInitAi result
salInitAi(...) != 0 emits cmpwi (signed) where target uses cmplwi; the
sibling salInitDsp already had a (u32) cast, salInitAi was missing it.
Byte-exact.

+++ src/main/audio/hw_init.c
@@ -187,7 +187,7 @@ int hwInit(u32 *sampleRate, u8 valueA, u8 valueB, u32 flags)
     salAuxFrame = 0;
     salMessageCallback = 0;
 
-    if (salInitAi(snd_handle_irq, flags, sampleRate) != 0 &&
+    if ((u32)salInitAi(snd_handle_irq, flags, sampleRate) != 0 &&
         salInitDspCtrl(valueA, valueB, (flags & 1) != 0) != 0 &&
         (u32)salInitDsp(flags) != 0) {
         sndEnd();
